make grid use backbone models & collections
This contains preliminary work on the refactor. This code is not
completely there yet, but we need to focus on getting the server
parts up and running before this can get merged. All thats left
to do on this is getting pagination to work as well as going through
what remains of the mainAction method an ensuring that the feature
still works after the rewrite.