appdrawer for startpage
Changed everything to initialize based on data.json in the project
root. This now initializes the menu part as well as the app overview
on the welcome page. The MenuItem models have been moved to core
as the new App model system. As of now apps may either be placed on
the menu or on an panel on the welcome page.